+++ fjs/cas/cli/module.f.ts
@@ -5,7 +5,8 @@
  */
 import { sha256 } from '../../crypto/sha2/module.f.ts'
 import { cBase32ToVec, vecToCBase32 } from '../../basen/cbase32/module.f.ts'
-import { eff, forEachStep, pure } from '../../effects/module.f.ts'
+import { forEachStep, pure } from '../../effects/module.f.ts'
+import { eff } from '../../effects/eff/module.f.ts'
 import {
     errorExit,
     log,
