+++ fjs/cas/evo/module.f.ts
@@ -32,7 +32,8 @@
  *
  * @module
  */
-import { eff, pure, foldStep, type Effect, type Operation } from '../../effects/module.f.ts'
+import { pure, foldStep, type Effect, type Operation } from '../../effects/module.f.ts'
+import { eff } from '../../effects/eff/module.f.ts'
 import { create, read, write, type Key, type MemOp } from '../../effects/memory/module.f.ts'
 import { collectRead, type Cas } from '../module.f.ts'
 import { cBase32ToVec, vecToCBase32 } from '../../basen/cbase32/module.f.ts'
