+++ fjs/cas/module.f.ts
@@ -7,7 +7,8 @@ import { sha256, type Sha2, type State as Sha2State } from '../crypto/sha2/modul
 import { join, normalize, parse } from '../path/module.f.ts'
 import { empty, length, maxLength, maxLengthBytes, msb, vec, type Vec } from '../types/bit_vec/module.f.ts'
 import { cBase32ToVec, vecToCBase32 } from '../basen/cbase32/module.f.ts'
-import { eff, foldStep, forEachStep, okStep, pure, type Effect, type Operation } from '../effects/module.f.ts'
+import { foldStep, forEachStep, okStep, pure, type Effect, type Operation } from '../effects/module.f.ts'
+import { eff } from '../effects/eff/module.f.ts'
 import {
     access,
     createExclusive,
