+++ fjs/cli/module.f.ts
@@ -1,5 +1,6 @@
 import { errorExit, log, type NodeOp, type NodeProgramOptions, type Write } from '../effects/node/module.f.ts'
-import { eff, pure, type Effect } from '../effects/module.f.ts'
+import { pure, type Effect } from '../effects/module.f.ts'
+import { eff } from '../effects/eff/module.f.ts'
 import { at, fromEntries } from '../types/object/module.f.ts'
 
 type Handler<O extends NodeOp> = (options: NodeProgramOptions) => Effect<O, number>
