+++ fjs/mcp/module.f.ts
@@ -16,7 +16,8 @@
 import { boolean, string, option, array, record, or } from '../types/rtti/module.f.ts'
 import { unknown, type Unknown } from '../media/json/module.f.ts'
 import type { Ts } from '../types/rtti/ts/module.f.ts'
-import { eff, pure, type Operation, type Effect } from '../effects/module.f.ts'
+import { pure, type Operation, type Effect } from '../effects/module.f.ts'
+import { eff } from '../effects/eff/module.f.ts'
 import { read, write, type Key, type MemOp } from '../effects/memory/module.f.ts'
 import {
     decodeRequest,
