+++ fjs/website/module.f.ts
@@ -5,7 +5,8 @@
  */
 import { htmlUtf8 } from '../media/html/module.f.ts'
 import { writeFile, type WriteFile } from '../effects/node/module.f.ts'
-import { eff, pure, type Effect } from '../effects/module.f.ts'
+import { pure, type Effect } from '../effects/module.f.ts'
+import { eff } from '../effects/eff/module.f.ts'
 import type { Vec } from '../types/bit_vec/module.f.ts'
 
 const html: Vec = htmlUtf8()(
